+++ app/controllers/concerns/including.rb
@@ -1,11 +1,11 @@
 module Including
-  UNPERMITTED_FILTER_PARAMS = ['**'].freeze
+  UNPERMITTED_INCLUDE_PARAMS = ['**'].freeze
 
   private
 
   def include_params
     return [] unless params[:include]
-    params[:include].split(',') - UNPERMITTED_FILTER_PARAMS
+    params[:include].split(',') - UNPERMITTED_INCLUDE_PARAMS
   end
 
   def include_associations