+++ app/serializers/contact_serializer.rb
@@ -72,13 +72,4 @@ class ContactSerializer < ApplicationSerializer
   def account_list_id
     object.account_list.uuid
   end
-
-  def lifetime_donations
-    # This causes performance issues, it should not be added until a solution is found.
-    object.donations.sum(:amount)
-  end
-
-  def tag_list
-    object.tags.collect(&:name)
-  end
 end