+++ app/services/contact/filter/tags.rb
@@ -1,5 +1,5 @@
 class Contact::Filter::Tags < Contact::Filter::Base
   def execute_query(contacts, filters)
-    contacts.tagged_with(filters[:tags].split(',').map(&:strip), any: filters[:any_tags] == 'true')
+    contacts.tagged_with(filters[:tags].split(',').map(&:strip), any: filters[:any_tags].to_s == 'true')
   end
 end