+++ app/services/task/filter/tags.rb
@@ -1,5 +1,5 @@
 class Task::Filter::Tags < Task::Filter::Base
   def execute_query(tasks, filters)
-    tasks.tagged_with(filters[:tags].split(',').map(&:strip), any: filters[:any_tags] == 'true')
+    tasks.tagged_with(filters[:tags].split(',').map(&:strip), any: filters[:any_tags].to_s == 'true')
   end
 end