+++ app/services/tnt_import/contact_import.rb
@@ -70,12 +70,12 @@ class TntImport::ContactImport
     contact.name = row['FileAs'] if @override || contact.name.blank?
     contact.full_name = row['FullName'] if @override || contact.full_name.blank?
     contact.greeting = row['Greeting'] if @override || contact.greeting.blank?
+    contact.envelope_greeting = extract_envelope_greeting_from_row(row) if @override || contact.envelope_greeting.blank?
     contact.website = row['WebPage'] if @override || contact.website.blank?
     contact.updated_at = parse_date(row['LastEdit']) if @override
     contact.created_at = parse_date(row['CreatedDate']) if @override
 
-    contact.add_to_notes(row['Notes'])
-    contact.add_to_notes("Children: #{row['Children']}") if must_add_children?(row, contact)
+    add_notes(contact, row)
 
     contact.pledge_amount = row['PledgeAmount'] if @override || contact.pledge_amount.blank?
     # PledgeFrequencyID: Since TNT 3.2, a negative number indicates a fequency in days. For example: -11 would be a frequency of 11 days. For now we are ignoring negatives.
@@ -143,6 +143,29 @@ class TntImport::ContactImport
     company.save!
     donor_account.update_attribute(:master_company_id, company.master_company_id) unless donor_account.master_company_id == company.master_company.id
     company
+  end
+
+  def add_notes(contact, row)
+    contact.add_to_notes(row['Notes'])
+    contact.add_to_notes("Children: #{row['Children']}") if must_add_children?(row, contact)
+
+    # These social fields don't have equivalents in MPDX, so we'll add them to notes:
+
+    contact.add_to_notes("Other Social: #{row['SocialWeb4']}") if row['SocialWeb4']
+    contact.add_to_notes("Spouse Other Social: #{row['SpouseSocialWeb4']}") if row['SpouseSocialWeb4']
+
+    contact.add_to_notes("Voice/Skype: #{row['VoiceSkype']}") if row['VoiceSkype']
+    contact.add_to_notes("Spouse Voice/Skype: #{row['SpouseVoiceSkype']}") if row['SpouseVoiceSkype']
+
+    contact.add_to_notes("IM Address: #{row['IMAddress']}") if row['IMAddress']
+    contact.add_to_notes("Spouse IM Address: #{row['SpouseIMAddress']}") if row['SpouseIMAddress']
+  end
+
+  def extract_envelope_greeting_from_row(row)
+    # TNT has something called a "MailingAddressBlock", the envelope greeting is the first line of this string.
+    block = row['MailingAddressBlock']
+    envelope_greeting = block.split("\n").detect(&:present?) # Find the first non-blank line of the string.
+    envelope_greeting.presence || row['FullName']
   end
 
   def true?(val)