+++ spec/controllers/api/v2/contacts/exports_controller_spec.rb
@@ -9,8 +9,8 @@ describe Api::V2::Contacts::ExportsController, type: :controller do
   let(:account_list) { user.account_lists.first }
   let(:account_list_id) { account_list.uuid }
 
-  let!(:contact) { create(:contact, account_list: account_list) }
-  let!(:second_contact) { create(:contact, account_list: account_list) }
+  let!(:contact) { create(:contact, account_list: account_list, name: 'Last Contact') }
+  let!(:second_contact) { create(:contact, account_list: account_list, name: 'First Contact') }
 
   let(:id) { contact.uuid }
 
@@ -26,38 +26,45 @@ describe Api::V2::Contacts::ExportsController, type: :controller do
   end
 
   context 'CSV export' do
-    it 'renders the export for users that are signed in' do
+    let(:contact_index) { response.body.index(contact.name) }
+    let(:second_contact_index) { response.body.index(second_contact.name) }
+
+    it 'renders the export sorted alphabetically for users that are signed in' do
       api_login(user)
       get :index, format: :csv
       expect(response.status).to eq(200)
-      expect(response.body).to include(contact.name)
-      expect(response.body).to include(second_contact.name)
+      expect(contact_index).to be_present
+      expect(response.body).to be_present
+      expect(contact_index).to be > second_contact_index
     end
 
     it 'renders the export with right contacts when contact_ids is provided' do
       api_login(user)
       get :index, format: :csv, filter: { ids: contact.uuid }
       expect(response.status).to eq(200)
-      expect(response.body).to include(contact.name)
-      expect(response.body).to_not include(second_contact.name)
+      expect(contact_index).to be_present
+      expect(response.body).to be_present
     end
   end
 
   context 'XLSX export' do
-    it 'renders the export for users that are signed in' do
+    let(:contact_index) { spreadsheet.to_csv.index(contact.name) }
+    let(:second_contact_index) { spreadsheet.to_csv.index(second_contact.name) }
+
+    it 'renders the export sorted alphabetically for users that are signed in' do
       api_login(user)
       get :index, format: :xlsx
-      expect(response.status).to eq(200)
-      expect(spreadsheet.to_csv).to include(contact.name)
-      expect(spreadsheet.to_csv).to include(second_contact.name)
+      expect(contact_index).to be_present
+      expect(second_contact_index).to be_present
+      expect(contact_index).to be > second_contact_index
     end
 
     it 'renders the export with right contacts when contact_ids is provided' do
       api_login(user)
       get :index, format: :xlsx, filter: { ids: contact.uuid }
       expect(response.status).to eq(200)
-      expect(spreadsheet.to_csv).to include(contact.name)
-      expect(spreadsheet.to_csv).to_not include(second_contact.name)
+      expect(contact_index).to be_present
+      expect(second_contact_index).to be_blank
     end
   end