+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This changelog covers what's changed in the MPDX APIs.
 
+## 5 April 2017
+- Added total_donation to donation detail
+- Amount under donation endpoint is now just returning an amount without symbol
+
+## 4 April 2017
+- TNT Import now imports social media accounts.
+- TNT Import now imports contact envelope_greeting.
+
 ## 3 April 2017
 - Imports API now accepts comma delimited tag_list, instead of tags.